requirements.txt
typer==0.6.1
pytorch==2.2.0
